+++ types/index.d.ts
@@ -107,9 +107,17 @@ declare namespace PhonegapPluginPush {
 		/**
 		 * Tells the OS to clear all notifications from the Notification Center
 		 * @param successHandler Is called when the api successfully clears the notifications.
-		 * @param errorHandler Is called when the api encounters an error when attempting to clears the notifications.
+		 * @param errorHandler Is called when the api encounters an error when attempting to clear the notifications.
 		 */
 		clearAllNotifications(successHandler: () => any, errorHandler: () => any): void
+
+		/**
+		 * Tells the OS to clear the notification that corresponds to the id argument, from the Notification Center
+		 * @param successHandler Is called when the api successfully clears the notification.
+		 * @param errorHandler Is called when the api encounters an error when attempting to clear the notification.
+		 * @param id The ID of the notification that will be cleared.
+		 */
+		clearNotification(successHandler: () => any, errorHandler: () => any, id?: number): void
 	}
 
 	/**
@@ -221,9 +229,13 @@ declare namespace PhonegapPluginPush {
 			 */
 			categories?: CategoryArray
 			/**
-			 * Whether to use prod or sandbox GCM setting. Defaults to false.
+			 * If `true` the device can show up critical alerts. (Possible since iOS 12 with a special entitlement)
+			 * Default is false|"false".
+       * Note: the value you set this option to the first time you call the init method will be how the application always acts.
+       * Once this is set programmatically in the init method it can only be changed manually by the user in Settings > Notifications > `App Name`.
+       * This is normal iOS behaviour.
 			 */
-			fcmSandbox?: boolean
+			critical?: boolean
 			/**
 			 * If the array contains one or more strings each string will be used to subscribe to a FcmPubSub topic. Defaults to [].
 			 */